+++ CHANGELOG.md
@@ -17,6 +17,7 @@ UNRELEASED
 * Improve performance of content forms
 * Improve performance of translation coverage view
 * Support legacy XLIFF export for MemoQ WPML filter
+* Improve performance of content searches
 
 
 2021.12.0-beta